Move subTopicIds function to articleUtils and import in subjectArticleList service MIM-2226
